Cleaned code. Changed data from .js to .json. Refactored logic for About and Portfolio.js. Improved and nitpicked style
