+++ src/content/about-content.json
@@ -1,0 +1,28 @@
+{
+  "name": "Christian Palacios",
+  "role": "Full Stack Web Developer",
+  "title": "About Me",
+  "paragraphs": [
+    {
+      "id": 1,
+      "text": "Hello there! I'm Christian Palacios, a 26-year-old full stack web developer, and I'm thrilled to share a bit about myself with you. From my passion for coding to my love for nature, music, and cars, I believe life is all about the journey, and I'd love for you to join me on mine."
+    },
+    {
+      "id": 2,
+      "text": "I'm not just a full stack web developer; I'm a creator, an artist, and a problem solver all rolled into one. I specialize in a variety of web technologies, including HTML, CSS, JavaScript, jQuery, Express.js, Node.js, SQL, Sequelize, MongoDB, Mongoose, React, PWAs, and GraphQL. For me, coding is where creativity meets logic, and each project is like a puzzle waiting to be solved. This is my true calling, and I can't wait to continue crafting digital magic."
+    },
+    {
+      "id": 3,
+      "text": "Before coding, I worked as an automotive technician, service advisor, service admin. I have a knack for taking things apart, putting them back together, and making them better. Whether it's a complex web application or my own car, I enjoy fixing things. Speaking of cars, I recently installed five dash cams on my vehicle because, well, you can never be too safe on the road! (if you're curious, check out my TikTok)."
+    },
+    {
+      "id": 4,
+      "text": "When I'm not immersed in the world of web development, you can find me out and about in nature. I have a deep appreciation for the great outdoors and a special fondness for animals. Whether it's a day at the park, a peaceful lakeside retreat, or a weekend camping adventure, I'm in my element. Music is another passion of mine. You might even catch me playing the piano on occasion, and if you're curious, check out my TikTok! To keep my body and mind in balance, I also enjoy working out, with running being a favorite."
+    },
+    {
+      "id": 5,
+      "text": "For me, the most important aspect of life is the journey of personal growth and learning. It's about developing one's character and pushing past mental limits. To achieve this, I believe in nurturing the mind, body, emotions, and spirit. It all starts with finding your purpose, and for me, my purpose is deeply rooted in coding. Join me as I continue on this incredible journey of self-discovery and web development!"
+    }
+  ],
+  "slogan": "Follow Your Heart!"
+}
