--- src/content/portfolioContent.js
@@ -1,65 +1,0 @@
-const portfolioContent = {
-  title: "Portfolio",
-  projects: [
-    {
-      id: 1,
-      title: "Unity Biz",
-      imageUrl: "/src/assets/projects-photos/project-unitybiz.png",
-      deployedLink: "https://unitybiz-fd2efb602da8.herokuapp.com/",
-      githubRepo: "https://github.com/rune-encoder/UnityBiz",
-    },
-    {
-      id: 2,
-      title: "Ultimate Tech Blog",
-      imageUrl: "/src/assets/projects-photos/project-tech-blog.png",
-      deployedLink: "https://u-tech-blog-84c6469a8f10.herokuapp.com/",
-      githubRepo: "https://github.com/rune-encoder/Ultimate-Tech-Blog",
-    },
-    {
-      id: 3,
-      title: "Note Logger",
-      imageUrl: "/src/assets/projects-photos/project-note-logger.png",
-      deployedLink: "https://note-logger-03a612b15e73.herokuapp.com/",
-      githubRepo: "https://github.com/rune-encoder/Note-Logger",
-    },
-    {
-      id: 4,
-      title: "Vroomu: A Vehicle Service Guide",
-      imageUrl: "/src/assets/projects-photos/project-vroomu.png",
-      deployedLink:
-        "https://gera1313.github.io/Vroomu-A-Vehicle-Service-Guide/",
-      githubRepo:
-        "https://github.com/rune-encoder/Vroomu-A-Vehicle-Service-Guide",
-    },
-    {
-      id: 5,
-      title: "Weather Dashboard",
-      imageUrl: "/src/assets/projects-photos/project-weather-dashboard.png",
-      deployedLink: "https://rune-encoder.github.io/Weather-Dashboard/",
-      githubRepo: "https://github.com/rune-encoder/Weather-Dashboard",
-    },
-    {
-      id: 6,
-      title: "Work Day Scheduler",
-      imageUrl: "/src/assets/projects-photos/project-work-day.png",
-      deployedLink: "https://rune-encoder.github.io/Work-Day-Scheduler/",
-      githubRepo: "https://github.com/rune-encoder/Work-Day-Scheduler",
-    },
-    {
-      id: 7,
-      title: "Code Quiz",
-      imageUrl: "/src/assets/projects-photos/project-code-quiz.png",
-      deployedLink: "https://rune-encoder.github.io/Code-Quiz/",
-      githubRepo: "https://github.com/rune-encoder/Code-Quiz",
-    },
-    {
-      id: 8,
-      title: "Password Generator",
-      imageUrl: "/src/assets/projects-photos/project-password-generator.png",
-      deployedLink: "https://rune-encoder.github.io/Password-Generator/",
-      githubRepo: "https://github.com/rune-encoder/Password-Generator",
-    },
-  ],
-};
-
-export default portfolioContent;
