+++ src/pages/About.jsx
@@ -1,17 +1,21 @@
-import aboutContent from "../content/aboutContent.js";
+import aboutContent from "../content/about-content.json";
 
 import profilePhoto from "../assets/profile-photo.jpg";
 import "../styles/About.css";
 
 export default function About() {
-  const { title, paragraphs, slogan } = aboutContent;
+  const { title, paragraphs, slogan, name, role } = aboutContent;
 
   return (
-    <div className="about-container grid gap-4 laptop:grid-cols-3 large-mobile:grid-cols-2 mobile:grid-cols-1">
+    <div className="about-container grid gap-4 laptop:grid-cols-3 large-mobile:grid-cols-1">
       <div className="image-group laptop:w-auto">
         <img className="profile-photo" src={profilePhoto} alt="Profile" />
         <div className="profile-desc">
-          <p>Christian Palacios | Full-Stack Web Developer.</p>
+          <p>
+            {name}
+            <br></br>
+            {role}
+          </p>
         </div>
       </div>
       <div className="about-group laptop:col-span-2">
