+++ src/styles/Footer.css
@@ -75,6 +75,7 @@ footer {
 
 .footer-text {
   margin-top: 5px;
+  color: var(--always-light);
   filter: drop-shadow(5px 3px 10px var(--opacity-medium));
 
   /*! Revisit */
